requirements.txt
python-dotenv==1.2.1
langgraph==1.0.8
langchain-google-genai==4.2.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
uvicorn==0.40.0
fastapi==0.128.6
streamlit==1.54.0
pandas==2.3.3
